requirements.txt
pandas>=2.2.0
numpy>=1.26.0
duckdb>=1.1.0
scikit-learn>=1.4.0
xgboost>=2.0.0
imbalanced-learn>=0.12.0
shap>=0.45.0
optuna>=3.6.0
python-dotenv>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
pyarrow>=16.0.0
joblib>=1.4.0
jupyter>=1.0.0
ipykernel>=6.29.0
